Add more naturality and randomness to the algo
Now allows the highest of highs in resolution by making up for the distance between the parts by tweening a larger part from the last part to the current part. This also clones the part every wait(resolution/10) and adds a light inside of the clone to light up the cave, but also not make it too bright. Resolution makeup also uses a random BALLSIZE variable when filling terrain, which I personally thought added more naturality. The last file algorithm just did not look right.